dataflow-analytics/web_interface/requirements.txt
flask==3.0.3
requests==2.25.1
PyYAML==5.4.1
werkzeug==2.0.3
